Normalize IV from protection header
The IV from the protection header may be less than 16 bytes long. In that case we pad it with leading zeros to make it 16 bytes. This makes the IV match the format that the decryption logic expects.

Tested with a stream using IV=1 (single byte)
http://playertest.longtailvideo.com/adaptive/oceans_aes/oceans_aes.m3u8
and some non-public local streams with 7 hex characters / 4 byte in key tag